requirements.txt
aiohappyeyeballs>=2.6.1
aiohttp>=3.13.2
aiosignal>=1.4.0
alembic>=1.17.2
annotated-doc>=0.0.4
annotated-types>=0.7.0
anyio>=4.12.0
asyncpg>=0.31.0
attrs>=25.4.0
cachetools>=6.2.4
certifi>=2025.11.12
charset-normalizer>=3.4.4
click>=8.3.1
colorama>=0.4.6
dataclasses-json>=0.6.7
distro>=1.9.0
fastapi>=0.128.0
filetype>=1.2.0
finnhub-python>=2.4.26
frozenlist>=1.8.0
google-auth>=2.45.0
google-genai>=1.56.0
greenlet>=3.3.0
h11>=0.16.0
httpcore>=1.0.9
httptools>=0.7.1
httpx>=0.28.1
httpx-sse>=0.4.3
idna>=3.11
jsonpatch>=1.33
jsonpointer>=3.0.0
langchain>=1.2.0
langchain-classic>=1.0.1
langchain-community>=0.4.1
langchain-core>=1.2.5
langchain-google-genai>=4.1.2
langchain-text-splitters>=1.1.0
langgraph>=1.0.5
langgraph-checkpoint>=3.0.1
langgraph-prebuilt>=1.0.5
langgraph-sdk>=0.3.1
langsmith>=0.5.1
Mako>=1.3.10
MarkupSafe>=3.0.3
marshmallow>=3.26.2
multidict>=6.7.0
mypy_extensions>=1.1.0
numpy>=2.4.0
orjson>=3.11.5
ormsgpack>=1.12.1
packaging>=25.0
pgvector>=0.4.2
propcache>=0.4.1
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pydantic>=2.12.5
pydantic-settings>=2.12.0
pydantic_core>=2.41.5
python-dotenv>=1.2.1
PyYAML>=6.0.3
requests>=2.32.5
requests-toolbelt>=1.0.0
rsa>=4.9.1
sniffio>=1.3.1
SQLAlchemy>=2.0.45
starlette>=0.50.0
tenacity>=9.1.2
typing-inspect>=0.9.0
typing-inspection>=0.4.2
typing_extensions>=4.15.0
urllib3>=2.6.2
uuid_utils>=0.12.0
uvicorn>=0.38.0
watchfiles>=1.1.1
websockets>=15.0.1
xxhash>=3.6.0
yarl>=1.22.0
zstandard>=0.25.0
